refactor(#1002): switch runtime/adapters to lib/runtime/ modules (with shared pending-focus)
Replace inverted imports of components-v2/wiring/* in lib/runtime/adapters/ with
canonical lib/runtime/ modules. Safe now that #1044 (shared pending-focus.ts) is merged.
