+++ frontend/src/lib/runtime/adapters/vfo-adapter.ts
@@ -5,8 +5,8 @@
  */
 
 import { runtime } from '../frontend-runtime';
-import { toVfoProps, toVfoOpsProps } from '../../../components-v2/wiring/state-adapter';
-import type { VfoStateProps, VfoOpsProps } from '../../../components-v2/wiring/state-adapter';
+import { toVfoProps, toVfoOpsProps } from '../props/panel-props';
+import type { VfoStateProps, VfoOpsProps } from '../props/panel-props';
 
 export function deriveMainVfo(): VfoStateProps {
   return toVfoProps(runtime.state, 'main');
